backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-dateutil
pyodbc
pydantic
python-multipart